server/requirements.txt
flask
flask-jwt-extended
psycopg2-binary
flask-cors
bcrypt
